# Reduce regexp memory use in tracestate