--- a/trace/tracestate_test.go
+++ b/trace/tracestate_test.go
@@ -554,11 +554,35 @@ func TestTraceStateImmutable(t *testing.T) {
 }
 
 func BenchmarkParseTraceState(b *testing.B) {
-	traceStateStr := "somewhatRealisticKeyLength=someValueAbcdefgh1234567890"
-	b.ReportAllocs()
-	b.ResetTimer()
+	benches := []struct {
+		name string
+		in   string
+	}{
+		{
+			name: "single key",
+			in:   "somewhatRealisticKeyLength=someValueAbcdefgh1234567890",
+		},
+		{
+			name: "tenant single key",
+			in:   "somewhatRealisticKeyLength@someTenant=someValueAbcdefgh1234567890",
+		},
+		{
+			name: "three keys",
+			in:   "someKeyName.One=someValue1,someKeyName.Two=someValue2,someKeyName.Three=someValue3",
+		},
+		{
+			name: "tenant three keys",
+			in:   "someKeyName.One@tenant=someValue1,someKeyName.Two@tenant=someValue2,someKeyName.Three@tenant=someValue3",
+		},
+	}
+	for _, bench := range benches {
+		b.Run(bench.name, func(b *testing.B) {
+			b.ReportAllocs()
+			b.ResetTimer()
 
-	for i := 0; i < b.N; i++ {
-		_, _ = ParseTraceState(traceStateStr)
+			for i := 0; i < b.N; i++ {
+				_, _ = ParseTraceState(bench.in)
+			}
+		})
 	}
 }