@dmathieu 2023-10-23
Could you run benchmark tests (maybe we need a new benchmark?) to show the actual memory difference?