@johnbley 2023-10-23
> Could you run benchmark tests (maybe we need a new benchmark?) to show the actual memory difference?

Sure.  I loaded the relevant `var`s and `const`s into an otherwise empty program, ran with `GODEBUG=allocfreetrace=1` and saw the total allocated bytes (via some cli munging) drop from 1119248 to 226312 for a savings of ~892kb.