@johnbley 2023-10-24
I added a simple benchmark and it shows a reduction in runtime from the old code:
Old:
```
BenchmarkParseTraceState-16       	  844042	      1446 ns/op	     324 B/op	       6 allocs/op
```
New:
```
BenchmarkParseTraceState-16       	 1000000	      1030 ns/op	     322 B/op	       6 allocs/op
```
though of course the reduction in resident RAM usage is not captured in this.